p-payment/requirements.txt
playwright==1.51
requests
schedule
python-dotenv
httpx